+++ R/sequential_transformations.R
@@ -1,20 +1,23 @@
 #' Find and classify outliers functional outliers using Sequential Transformation
 #'
-#' This method finds and classify outliers using sequential transformations proposed in Algorithm 1 of
-#' Dai et al. (2020)\href{https://doi.org/10.1016/j.csda.2020.106960}{<doi:10.1016/j.csda.2020.106960>}.
-#' A sequence of transformations are applied to the functional data and after each transformation, a functional
-#' boxplot is applied on the transformed data and outliers flagged by the functional data are noted. A number of
-#' transformations mentioned in Dai et al. (2020)\href{https://doi.org/10.1016/j.csda.2020.106960}{<doi:10.1016/j.csda.2020.106960>}
-#' are supported including vertical alignment ("T1(X)(t)"), normalization ("T2(X)(t)"),
-#' one order of differencing ("D1(X)(t)" and "D2(X)(t)") and point-wise outlyingness data ("O(X)(t)"). The feature alignment
-#' transformation based on warping/curve registration is not yet supported.
+#' This method finds and classify outliers using sequential transformations
+#' proposed in Algorithm 1 of Dai et al. (2020) \doi{10.1016/j.csda.2020.106960}.
+#' A sequence of transformations are applied to the functional data and after each
+#' transformation, a functional boxplot is applied on the transformed data and outliers
+#' flagged by the functional data are noted. A number of transformations mentioned in
+#' Dai et al. (2020) \doi{10.1016/j.csda.2020.106960} are supported including vertical
+#' alignment ("T1(X)(t)"), normalization ("T2(X)(t)"), one order of differencing
+#' ("D1(X)(t)" and "D2(X)(t)") and point-wise outlyingness data ("O(X)(t)").
+#' The feature alignment transformation based on warping/curve registration is not yet
+#' supported.
 #'
 #'
-#' @param dt A matrix for univariate functional data (of size \code{n} observations by \code{p} domain
+#' @param dts A matrix for univariate functional data (of size \code{n} observations by \code{p} domain
 #'  points) or a 3-dimensional array for multivariate functional data (of size \code{n}
 #'  observations by \code{p} domain points by \code{d} dimension). Only the outlyingness transformation ("O(X)(t)")
 #'  supports multivariate functional data so the sequence of transformation must always start with outlyingness ("O(X)(t)")
-#'  whenever a multivariate functional data is parsed to \code{dt}.
+#'  whenever a multivariate functional data is parsed to \code{dts}.
+#'
 #' @param sequence A character vector usually of length between 1 and 6 containing any of the strings: \code{"T0", "D0", "T1", "T2",
 #' "D1", "D2"} and \code{"O"} (in any order). These sequence of strings specifies the sequence of transformations to be applied
 #'  on the data and their meanings are described as follows:
@@ -59,7 +62,7 @@
 #'  of upper and lower quantiles. Defaults to \code{c(0.025, 0.975)} for the upper and lower 2.5\% quantiles.
 #'  See \link{directional_quantile} for details.
 #' @param n_projections An integer indicating the number of random projections to use in computing the point-wise outlyingness if a 3-d array
-#'  is specified in \code{dt} i.e. (multivariate functional data), and the transformation \code{"O"} is part of the sequence of transformations
+#'  is specified in \code{dts} i.e. (multivariate functional data), and the transformation \code{"O"} is part of the sequence of transformations
 #'  parsed to \code{sequence}. Defaults to 200L.
 #' @param seed The random seed to set when generating the random directions in the computation of the point-wise outlyingness. Defaults to NULL.
 #' in which case a seed is not set.
@@ -72,12 +75,13 @@
 #'  \code{save_data = FALSE}).}
 #'
 #' @details
-#'This function implements outlier detection using sequential transformations described in Algorithm 1 of
-#'Dai et al. (2020) \href{https://doi.org/10.1016/j.csda.2020.106960}{<doi:10.1016/j.csda.2020.106960>}.
-#'A sequence of transformations are applied consecutively with the functional boxplot applied on the transformed data after
-#'each transformation. The following example sequences (and their meaning) suggested in
-#'Dai et al. (2020) \href{https://doi.org/10.1016/j.csda.2020.106960}{<doi:10.1016/j.csda.2020.106960>} can be parsed to
-#'argument \code{sequence}.
+#'This function implements outlier detection using sequential transformations
+#'described in Algorithm 1 of Dai et al. (2020) \doi{10.1016/j.csda.2020.106960}.
+#'A sequence of transformations are applied consecutively with the functional
+#'boxplot applied on the transformed data after each transformation. The following
+#'example sequences (and their meaning) suggested in Dai et al. (2020)
+#'\doi{10.1016/j.csda.2020.106960} can be parsed to argument \code{sequence}.
+#'
 #'\describe{
 #'   \item{\code{"T0"}}{Apply functional boxplot on raw data (no transformation is applied).}
 #'    \item{\code{c("T0", "T1", "D1")}}{Apply functional boxplot on raw data, then apply vertical alignment on data followed by applying
@@ -93,7 +97,7 @@
 #'    "one_sided_right" extreme rank length depth) in the functional boxplot used on the data of point-wise outlyingness. This is because only
 #'    large values should be considered extreme in the data of the point-wise outlyingness.}
 #'  }
-#' For multivariate functional data (when a 3-d array is supplied to \code{dt}), the sequence of transformation must always begin with \code{"O"}
+#' For multivariate functional data (when a 3-d array is supplied to \code{dts}), the sequence of transformation must always begin with \code{"O"}
 #' so that the multivariate data can be replaced with the univariate data of point-wise outlyingness which the functional boxplot can subsequently process
 #' because the \code{\link{functional_boxplot}} function only supports univariate functional data.
 #'
@@ -106,18 +110,20 @@
 #'
 #' @examples
 #' # same as running a functional boxplot
-#' seqobj <- seq_transform(sim_data1$data, sequence = "T0", depth_method = "mbd")
+#' dt1 <- simulation_model1()
+#' seqobj <- seq_transform(dt1$data, sequence = "T0", depth_method = "mbd")
 #' seqobj$outliers$T0
-#' functional_boxplot(sim_data1$data, depth_method = "mbd")$outliers
+#' functional_boxplot(dt1$data, depth_method = "mbd")$outliers
 #'
 #' # more sequences
-#' seqobj <- seq_transform(sim_data1$data, sequence = c("T0", "D1", "D2"), depth_method = "mbd")
+#' dt4 <- simulation_model4()
+#' seqobj <- seq_transform(dt4$data, sequence = c("T0", "D1", "D2"), depth_method = "mbd")
 #' seqobj$outliers$T0 # outliers found in raw data
 #' seqobj$outliers$D1 # outliers found after differencing data the first time
 #' seqobj$outliers$D2 # outliers found after differencing the data the second time
 #'
 #' # saving transformed data
-#' seqobj <- seq_transform(sim_data1$data, sequence = c("T0", "D1", "D2"),
+#' seqobj <- seq_transform(dt4$data, sequence = c("T0", "D1", "D2"),
 #'  depth_method = "mbd", save_data = TRUE)
 #' seqobj$outliers$T0 # outliers found in raw data
 #' head(seqobj$transformed_data$T0)  # the raw data
@@ -125,7 +131,7 @@
 #' head(seqobj$transformed_data$D2) # the 2nd order differenced data
 #'
 #' # double transforms e.g. c("T0", "D1", "D1")
-#' seqobj <- seq_transform(sim_data1$data, sequence = c("T0", "D1", "D1"),
+#' seqobj <- seq_transform(dt4$data, sequence = c("T0", "D1", "D1"),
 #'  depth_method = "mbd", save_data = TRUE) # throws warning
 #' seqobj$outliers$T0 # outliers found in raw data
 #' seqobj$outliers$D1_1 #found after differencing data the first time
@@ -135,16 +141,16 @@
 #' head(seqobj$transformed_data$D1_2) # the 2nd order differenced data
 #'
 #' # multivariate data
-#' dtm <- array(0, dim = c(100, 50, 2))
-#' dtm[,,1] <- sim_data1$data
-#' dtm[,,2] <- sim_data1$data
+#' dtm <- array(0, dim = c(dim(dt1$data), 2))
+#' dtm[,,1] <- dt1$data
+#' dtm[,,2] <- dt1$data
 #' seqobj <- seq_transform(dtm, sequence = "O", depth_method = "erld",
 #'  erld_type = "one_sided_right", save_data = TRUE)
 #' seqobj$outliers$O # multivariate outliers
 #' head(seqobj$transformed_data$O) # univariate outlyingness data
 #'
 
-seq_transform <- function(dt, sequence = c("T0", "T1", "T2"),
+seq_transform <- function(dts, sequence = c("T0", "T1", "T2"),
                           depth_method = c("mbd", "tvd", "extremal", "dirout",
                                     "linfinity", "bd", "erld", "dq"),
                           save_data = FALSE,
@@ -154,6 +160,7 @@ seq_transform <- function(dt, sequence = c("T0", "T1", "T2"),
                           dq_quantiles = NULL,
                           n_projections = 200L,
                           seed = NULL){
+
   outliers <- list()
   if(save_data){
     transformed_data <- list()
@@ -164,50 +171,50 @@ seq_transform <- function(dt, sequence = c("T0", "T1", "T2"),
     transformation = sequence[i]
     if (transformation == "T0" || transformation == "D0"){
       ## apply functional boxplot here
-      t0_outliers <- functional_boxplot(dt, depth_method = depth_method,
+      t0_outliers <- functional_boxplot(dts, depth_method = depth_method,
                                         central_region = central_region,
                                         emp_factor = emp_factor,
                                         erld_type = erld_type,
                                         dq_quantiles = dq_quantiles)$outliers
       outliers[[i]] <- t0_outliers
-      if(save_data) transformed_data[[i]] <- dt
+      if(save_data) transformed_data[[i]] <- dts
     }else if(transformation == "T1"){
-      dt <- center_curves(dt)
-      t1_outliers <- functional_boxplot(dt, depth_method = depth_method,
+      dts <- center_curves(dts)
+      t1_outliers <- functional_boxplot(dts, depth_method = depth_method,
                                         emp_factor = emp_factor,
                                         central_region = central_region,
                                         erld_type = erld_type,
                                         dq_quantiles = dq_quantiles)$outliers
       outliers[[i]] <- t1_outliers
-      if(save_data) transformed_data[[i]] <- dt
+      if(save_data) transformed_data[[i]] <- dts
     } else if(transformation == "T2"){
-      dt <- normalize_curves(dt)
-      t2_outliers <- functional_boxplot(dt, depth_method = depth_method,
+      dts <- normalize_curves(dts)
+      t2_outliers <- functional_boxplot(dts, depth_method = depth_method,
                                         emp_factor = emp_factor,
                                         central_region = central_region,
                                         erld_type = erld_type,
                                         dq_quantiles = dq_quantiles)$outliers
       outliers[[i]] <- t2_outliers
-      if(save_data) transformed_data[[i]] <- dt
+      if(save_data) transformed_data[[i]] <- dts
     } else if(transformation == "D1"|| transformation == "D2"){
-      dt <- difference_curves(dt)
-      d1_outliers <- functional_boxplot(dt, depth_method = depth_method,
+      dts <- difference_curves(dts)
+      d1_outliers <- functional_boxplot(dts, depth_method = depth_method,
                                         emp_factor = emp_factor,
                                         central_region = central_region,
                                         erld_type = erld_type,
                                         dq_quantiles = dq_quantiles)$outliers
       outliers[[i]] <- d1_outliers
-      if(save_data) transformed_data[[i]] <- dt
+      if(save_data) transformed_data[[i]] <- dts
     } else if(transformation == "O"){
       # implement directional quantile!
-      dt <- outlyingness_curves(dt,  n_projections = n_projections, seed = seed)
-      o_outliers <- functional_boxplot(dt, depth_method = depth_method,
+      dts <- outlyingness_curves(dts,  n_projections = n_projections, seed = seed)
+      o_outliers <- functional_boxplot(dts, depth_method = depth_method,
                                        emp_factor = emp_factor,
                                        central_region = central_region,
                                        erld_type = erld_type,
                                        dq_quantiles = dq_quantiles)$outliers
       outliers[[i]] <- o_outliers
-      if(save_data) transformed_data[[i]] <- dt
+      if(save_data) transformed_data[[i]] <- dts
 
     }else {
       stop("Transformation ", transformation, ' not supported. \n')