+++ NEWS.md
@@ -1,3 +1,11 @@
-# fdaoutlier 0.1.0
+# fdaoutlier 0.2.0
+* Add msplot plotting function
+* Access outliers in msplot with `msplot_object$outliers` instead of `msplot_outlier$outliers_index`.
+* Add nine simulation models with plotting function.
+* Add MUOD method.
 
+# fdaoutlier 0.1.1
+* patch for error on CRAN checks
+
+# fdaoutlier 0.1.0
 * `fdaoutlier` Submitted to CRAN